rename recipesSlice -> modDefinitionsSlice
